Complete package refactor
