+++ README.md
@@ -1,5 +1,5 @@
-# julia-gam
-Explores some generalised additive model (GAM) and related (polynomial spline) architecture by hand in Julia.
+# GAMsByHand
+Builds some generalised additive model (GAM) and related (polynomial basis spline) architecture from scratch in Julia.
 
 ## Notes
